ready for matrix calculations and more unit tests
