+++ package.json
@@ -13,6 +13,7 @@
     "axios": "^0.19.2",
     "cors": "^2.8.5",
     "express": "^4.17.1",
+    "ml-matrix": "^6.5.0",
     "pg": "^8.0.0"
   },
   "devDependencies": {
